Refactor startup into UseBareMetalWeb() extension method
Extract all boilerplate setup (data store, rendering, routing, CORS, HTTPS,
static files, proxy) into BareMetalWebExtensions.UseBareMetalWeb().

Program.cs now just calls UseBareMetalWeb() with a callback for app-specific
routes (device code auth, ideas/search, API metadata).

All standard routes (static, auth, monitoring, admin, data, API CRUD) are
registered by RouteRegistrationExtensions and no longer duplicated inline.

Closes #29
